pyproject.toml
[project]
name = "sitetrackerduplicatedetection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.34",
    "pillow>=10.3",
    "imagehash>=4.3",
    "numpy>=1.26",
    "pandas>=2.2",
    "pyarrow>=16.1",
    "sqlalchemy>=2.0",
    "pyyaml>=6.0",
    "psycopg2-binary>=2.9",
]

[tool.uv]
allow-insecure-host = ["pypi.org", "files.pythonhosted.org"]
